Move method to interface step 1
